Small optimisation in coreclient
